LPS-116770 Moved to utils the functions to check wheter an action can be executed or not
