+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/utils/canBeRemoved.js
@@ -1,0 +1,27 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+import hasDropZoneChild from '../components/layout-data-items/hasDropZoneChild';
+import {LAYOUT_DATA_ITEM_TYPES} from '../config/constants/layoutDataItemTypes';
+
+export default function canBeRemoved(item, layoutData) {
+	switch (item.type) {
+		case LAYOUT_DATA_ITEM_TYPES.column:
+		case LAYOUT_DATA_ITEM_TYPES.dropZone:
+			return false;
+
+		default:
+			return !hasDropZoneChild(item, layoutData);
+	}
+}
